requirements.txt
openpyxl==3.0.10
pandas==1.5.0
pytesseract==0.3.10
Flask==2.2.2
PyMuPDF==1.20.2
scipy==1.9.1
awscli==1.25.77
boto3==1.24.76
python-dotenv==0.21.0
requests==2.28.1
opencv-python==4.6.0.66
Flask-Cors==3.0.10
paho-mqtt==1.6.1
Flask-MQTT==1.1.1
Pillow==9.2.0
chardet==4.0.0